`pdflatex` alias is too important for this
I promoted this alias to a wrapper script in ~/bin, where it will
also be used when running pdflatex from vim.

Unfortunately this also has the potential to break software which
depends on pdflatex's default behavior, but that's a risk that I'm
honestly willing to take at this point.

Interative mode is *that bad*.