Limit memory in interactive shells.
In response to the following lovely shell session: (I was curious)

    $ python3
    Python 3.4.3 (default, Oct 14 2015, 20:28:29)
    [GCC 4.8.4] on linux
    Type "help", "copyright", "credits" or "license" for more information.
    >>> import itertools
    >>> a,b,*_ = itertools.count(0)
    ^C^C^C^C^C^C^C^C^C^C^C^C^C^C^C^C^C^C^C^C^C^C^C^C^C^C^C^C^C^C^C

Let's face it, *I'm an idiot*, and I am a danger to myself for as
long as I am capable of writing a script that can consume all of my
physical memory and begin disk thrashing faster than I can hit Ctrl+C.

I can think of nothing that I would ever run from within an interactive
shell session which would require this much memory.  Not even a tree of
processes. So if something uses this much memory, I can only imagine
that it MUST be, well, me being an idiot again.

Of course these kinds of things are preventable with care.
Hell, there's a module I wrote years ago for this exact purpose.
The only reason I didn't use it is because I would've had to write
two extra lines, which would require effort:

    >>> from pylibs.debug import memlimit
    >>> memlimit(100) # 100 MB
    Traceback (most recent call last):
      File "<stdin>", line 1, in <module>
      File "/home/lampam/.local/lib/python3.4/site-packages/pylibs/debug.py", line 3, in memlimit
        resource.setrlimit(resource.RLIMIT_AS, (megs * 1048576, -1))
    ValueError: not allowed to raise maximum limit

...Err, um, that's new.  That certainly isn't the output I expected
to receive and intended to show here.

Uh... fun times await?