Big backlog of uncommited changes
Some of these changes are fairly recent, like stuff for idris and
spacemacs.  Some are things that have been around but simply not in VCS.

Other things (in particular some changes in vim config) are
terrible, horrible sins against humanity that I never really
wanted to commit.
However, I figure it's best to get everything in sync before the
transfer to the new machine...