updates to git script
removes usage of git-worktree which is a recent feature

tries to catch other possible errors now, in the hopes that any
failures on future systems will simply result in that part of the
prompt disappearing, rather than uncontrollably spamming STDERR
and being a nuisance.

I also made the mistake of thinking to myself that I could reasonably
factor out a piece of shell code that needed to be sourced.
I REGRET EVERYTHING